[config] Disable futimens to fix backward deployment issue
This API is available in macOS 10.13, but since this configuration is
fixed, we should be conservative. In practice the only effect this has
is to fix a warning while building this code on macOS, because the only
function that uses futimens is setLastModificationAndAccessTime, which
is not used anywhere in indexstore-db.