[llvm] Mangle all llvm symbols to avoid collisions
Rename all of the llvm symbols so that they will not collide if another
copy of llvm is linked into the same image. The use of llvm within
IndexStore is purely an implementation detail. Using a source-level
rename is a workaround for the lack of symbol visibility controls in
swiftpm. Ideally we could do this with a combination of
`-fvisibility=hidden` and `ld -r`.

This allows us to link into sourcekit-lsp statically without colliding
with symbols in llbuild.