[IndexDatastore] For the out-of-date trigger notifications report all the out-of-date files without any coalescing
Previously if there were multiple headers that got out-of-date and they were all included by the same main file, it
would trigger a single notification for that main file. With these changes there will be notifications for each file.

This allows clients to get the full information of what needs to get updated and they get the responsibility for doing any
kind of coalescing they want.