CIndexStoreDB: correct dllstorage annotation
The `INDEXSTORE_PUBLIC` macro was guarding based on `_MSC_VER` which is
a check for MSVC's compiler which is incorrect.  This annotation has
nothing to do with the MSVC compiler but rather the PE/COFF file format.
Adjust that to `_WIN32` as an alias for Windows.

Furthermore, correct the annotation - the functions should be given
dllexport storage when building IndexStoreDB, and dllimport storage when
building code which uses IndexStoreDB (unless it is statically built).