[tsan] Fix a race during processUnitsAndWait()
After we added processUnitsAndWait() as a separte entry point that
ultimately called the onFilesChange() method, it created a race on the
DoneInitState, which was accessed directly in the unit event callback
before jumping to the unit processing queue. Simplify that logic and fix
the race by using processUnitsAndWait() during initialization if we
desire waiting, and get rid of DoneInitState.

The previous approach was more flexible about allowing you to wait at an
arbitrary point after creating the IndexDatastore, but in practice we
were not using that ability, and it made it harder to understand the
synchronization required.