Fix race in index delegate callback
All calls to `processingCompleted` are called from the unit processing
queue, but `processingAddedPending` can be called from other queues,
including an arbitrary queue in the case of `addUnitOutFilePaths` and
`removeUnitOutFilePaths`.

The result is that if there is a session being processed, we might
increment and decrement `PendingActions` concurrently. We obsevered an
assertion failure where `NumActions` was greater than `NumActions`,
which was likely caused by this race.

rdar://68024910