[tsan] Workaround uninstrumented library code
We get rare reports from TSan that there is a race between modifying the
UnitEventInfoDeque and its own destructor. This appears to be caused by
uninstrumented code in the indexstore library as well as the underlying
OS file watching library (FSEvents, inotify, etc.), which prevents TSan
from realizing that the event stream callback and the event stream
finalizer are never called concurrently. We workaround by introducing an
unnecessary mutex that we lock around the requisite calls.

rdar://69514830