[Index] Don't check a new unit for out-of-date dependencies, unless it is the initial scan
Out-of-date checks should happen for when doing the initial scan and when the file watcher finds that a dependency was modified.
Doing an out-of-date check for a newly created unit can lead to constant attempts at re-indexing an out-of-date unit, like
in the case where a dependency has a timestamp in the future, or if the dependency doesn't exist on disk for some reason.

But though the latter (missing dependency) is a problem, constantly sending out-of-date notifications for the unit is not helpful.

rdar://87127317