Add missing <ratio> header in "llvm/Support/Chrono.h"
`Chrono.h` uses `ratio`, but doesn't include its corresponding header.
This is fine on main and on the regular SPM build because `chrono`
includes `ratio` in one of its private headers and main's `chrono` just
exports all the headers. But on rebranch the private headers are now ...
private and thus `ratio` is no longer visible.

Include it directly in `Chrono.h` to unblock rebranch for now.