build: bump the C++ standard to 14
This matches the changes to Package.swift.  The use of the inline
initialised members for a struct in C++11 disqualify the type from being
an aggregate breaking the initialiser.  The semantics were changed in
C++14 which permit the construction of the `DelegateEvent`.  This
repairs the build with CMake and `clang++`.