Reduce memory consumption for OutOfDateTriggerHints
Remove `DependentUnitOutOfDateTriggerHint`, and
collapse the class hierarchy into a single class
`OutOfDateFileTrigger`, which stores both the file
name as well as the mod time. Then, change the
`OutOfDateTriggers` map in UnitMonitor to use
StringRef keys. This significantly reduces memory
consumption for large projects. We could likely
go further here by eliminating the storing of
triggers completely, but I'm leaving that as future
work for now.

rdar://50929788